init server timings
